requirements.txt
easydict==1.10
loguru==0.7.2
matplotlib==3.7.1
numpy==1.24.4
pandas==2.0.3
pymoo==0.5.0
PyYAML==6.0
PyYAML==6.0.1
scikit_learn==1.3.1
tabulate==0.9.0
thop==0.1.1.post2209072238
torch==2.0.0+cu118
tqdm==4.61.2
onnx==1.14.1